refactor: break up the kitchen sink class
